Add various enhancements like: non-altering-scaling, default font, and bugfixes.
This commit adds various fixes which at least locally, provide a better experience. This has been tested on FreeBSD13.3.

The option to scale the text of a terminal without altering the window dimensions has been added, with the command "smaller-static" and "bigger-static" used to ensure the dimensions of the terminal do not change.

In all cases, the location of the terminal is now retained: changing the size of the font -- whether it be with altering the window dimensions or not -- will no longer move the terminal's top-left x and y coordinates. This means the terminal will not only scale down and outwards (to the right).

In the case of no urxvt.font property being set, we fallback to the default: xft:monospace:size=8.

Finally, changed the documentation a little bit.